+++ Sources/Core/Anchorable.swift
@@ -9,7 +9,7 @@ import UIKit
 
 /// Provides useful features for anything that supports adding constraints to itself.
 /// Best usecases are Views and Guides.
-public protocol Anchorable {
+public protocol Anchorable: AnchorProvidable {
     /// Adds current view to provided parent view (if it's not already added)
     /// - Parameter parentView: a view on which current view (child view) needs to be added
     /// - Returns: current view
